Move RelatedFieldWidgetWrapper files to different directories
